+++ src/Subscriptions/StorageManager.php
@@ -24,11 +24,15 @@ class StorageManager implements StoresSubscriptions
     public const SUBSCRIBER_KEY = 'graphql.subscriber';
 
     /**
+     * The cache to store channels and topics.
+     *
      * @var \Illuminate\Contracts\Cache\Repository
      */
     protected $cache;
 
     /**
+     * The time to live for items in the cache.
+     *
      * @var \DateInterval|int|null
      */
     protected $ttl;
@@ -41,11 +45,6 @@ class StorageManager implements StoresSubscriptions
         $this->ttl = config('lighthouse.subscriptions.storage_ttl', null);
     }
 
-    /**
-     * Get subscriber by request.
-     *
-     * @return \Nuwave\Lighthouse\Subscriptions\Subscriber|null
-     */
     public function subscriberByRequest(array $input, array $headers): ?Subscriber
     {
         $channel = Arr::get($input, 'channel_name');
@@ -55,63 +54,113 @@ class StorageManager implements StoresSubscriptions
             : null;
     }
 
-    /**
-     * Find subscriber by channel.
-     *
-     * @return \Nuwave\Lighthouse\Subscriptions\Subscriber|null
-     */
     public function subscriberByChannel(string $channel): ?Subscriber
     {
-        $key = self::SUBSCRIBER_KEY.".{$channel}";
-
-        return $this->cache->get($key);
+        return $this->cache->get(self::channelKey($channel));
     }
 
-    /**
-     * Get collection of subscribers by channel.
-     *
-     * @return \Illuminate\Support\Collection<\Nuwave\Lighthouse\Subscriptions\Subscriber>
-     */
     public function subscribersByTopic(string $topic): Collection
     {
-        $channelsJson = $this->cache->get(self::TOPIC_KEY.".{$topic}");
-        if (! $channelsJson) {
-            return new Collection;
-        }
-
-        $channels = json_decode($channelsJson, true);
-
-        return (new Collection($channels))
+        return $this
+            ->retrieveTopic(self::topicKey($topic))
             ->map(function (string $channel): ?Subscriber {
                 return $this->subscriberByChannel($channel);
             })
-            ->filter()
-            ->values();
+            ->filter();
     }
 
     public function storeSubscriber(Subscriber $subscriber, string $topic): void
     {
-        $topicKey = self::TOPIC_KEY.".{$topic}";
-        $subscriberKey = self::SUBSCRIBER_KEY.".{$subscriber->channel}";
+        $subscriber->topic = $topic;
+        $this->addSubscriberToTopic($subscriber);
 
-        $topicJson = $this->cache->get($topicKey);
-        $topic = $topicJson
-            ? json_decode($topicJson, true)
-            : [];
-
-        $topic[] = $subscriber->channel;
-
+        $channelKey = self::channelKey($subscriber->channel);
         if ($this->ttl === null) {
-            $this->cache->forever($topicKey, json_encode($topic));
-            $this->cache->forever($subscriberKey, $subscriber);
+            $this->cache->forever($channelKey, $subscriber);
         } else {
-            $this->cache->put($topicKey, json_encode($topic), $this->ttl);
-            $this->cache->put($subscriberKey, $subscriber, $this->ttl);
+            $this->cache->put($channelKey, $subscriber, $this->ttl);
         }
     }
 
     public function deleteSubscriber(string $channel): ?Subscriber
     {
-        return $this->cache->pull(self::SUBSCRIBER_KEY.".{$channel}");
+        $subscriber = $this->cache->pull(self::channelKey($channel));
+
+        if ($subscriber !== null) {
+            $this->removeSubscriberFromTopic($subscriber);
+        }
+
+        return $subscriber;
+    }
+
+    /**
+     * Store a topic (list of channels) in the cache.
+     *
+     * @param  \Illuminate\Support\Collection<string>  $topic
+     */
+    protected function storeTopic(string $key, Collection $topic): void
+    {
+        if ($this->ttl === null) {
+            $this->cache->forever($key, $topic);
+        } else {
+            $this->cache->put($key, $topic, $this->ttl);
+        }
+    }
+
+    /**
+     * Retrieve a topic (list of channels) from the cache.
+     *
+     * @return \Illuminate\Support\Collection<string>
+     */
+    protected function retrieveTopic(string $key): Collection
+    {
+        return $this->cache->get($key, new Collection());
+    }
+
+    /**
+     * Add the subscriber to the topic they subscribe to.
+     *
+     * @param  \Nuwave\Lighthouse\Subscriptions\Subscriber  $subscriber
+     */
+    protected function addSubscriberToTopic(Subscriber $subscriber): void
+    {
+        $topicKey = self::topicKey($subscriber->topic);
+
+        $topic = $this->retrieveTopic($topicKey);
+        $topic->push($subscriber->channel);
+        $this->storeTopic($topicKey, $topic);
+    }
+
+    /**
+     * Remove the subscriber from the topic they are subscribed to.
+     */
+    protected function removeSubscriberFromTopic(Subscriber $subscriber)
+    {
+        $topicKey = self::topicKey($subscriber->topic);
+        $channelKeyToRemove = self::channelKey($subscriber->channel);
+
+        $topicWithoutSubscriber = $this
+            ->retrieveTopic($topicKey)
+            ->reject(function (string $channel) use ($channelKeyToRemove): bool {
+                return self::channelKey($channel) === $channelKeyToRemove;
+            });
+
+        if ($topicWithoutSubscriber->isEmpty()) {
+            $this->cache->forget($topicKey);
+
+            return;
+        }
+
+        $this->storeTopic($topicKey, $topicWithoutSubscriber);
+    }
+
+    protected static function channelKey(string $channel): string
+    {
+        return self::SUBSCRIBER_KEY.".{$channel}";
+    }
+
+    protected static function topicKey(string $topic): string
+    {
+        return self::TOPIC_KEY.".{$topic}";
     }
 }