+++ CHANGELOG.md
@@ -20,6 +20,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ### Changed
 
+- Remove subscriber reference from topic when deleted https://github.com/nuwave/lighthouse/pull/1288
 - Improve subscription context serializer https://github.com/nuwave/lighthouse/pull/1283
 - Allow replacing the `SubscriptionRegistry` implementation using the container https://github.com/nuwave/lighthouse/pull/1286